+++ app/src/main/java/com/teamwss/websoso/ui/mapper/NovelInfoMapper.kt
@@ -1,12 +1,12 @@
 package com.teamwss.websoso.ui.mapper
 
 import com.teamwss.websoso.data.model.NovelInfoEntity
-import com.teamwss.websoso.ui.novelDetail.novelInfo.model.KeywordModel
-import com.teamwss.websoso.ui.novelDetail.novelInfo.model.NovelInfoUiModel
-import com.teamwss.websoso.ui.novelDetail.novelInfo.model.PlatformModel
-import com.teamwss.websoso.ui.novelDetail.novelInfo.model.ReadStatus
-import com.teamwss.websoso.ui.novelDetail.novelInfo.model.ReviewCountModel
-import com.teamwss.websoso.ui.novelDetail.novelInfo.model.UnifiedReviewCountModel
+import com.teamwss.websoso.ui.novelInfo.model.KeywordModel
+import com.teamwss.websoso.ui.novelInfo.model.NovelInfoUiModel
+import com.teamwss.websoso.ui.novelInfo.model.PlatformModel
+import com.teamwss.websoso.ui.novelInfo.model.ReadStatus
+import com.teamwss.websoso.ui.novelInfo.model.ReviewCountModel
+import com.teamwss.websoso.ui.novelInfo.model.UnifiedReviewCountModel
 
 fun NovelInfoEntity.toUi() = NovelInfoUiModel(
     novelDescription = novelDescription,
