+++ app/src/main/java/com/teamwss/websoso/ui/novelDetail/adapter/NovelDetailPagerAdapter.kt
@@ -4,7 +4,7 @@ import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.teamwss.websoso.ui.novelDetail.NovelFeedFragment
-import com.teamwss.websoso.ui.novelDetail.novelInfo.NovelInfoFragment
+import com.teamwss.websoso.ui.novelInfo.NovelInfoFragment
 
 class NovelDetailPagerAdapter(fragmentActivity: FragmentActivity) :
     FragmentStateAdapter(fragmentActivity) {
