--- app/src/main/java/com/teamwss/websoso/ui/novelDetail/novelInfo/model/ReadStatus.kt
@@ -1,5 +1,0 @@
-package com.teamwss.websoso.ui.novelDetail.novelInfo.model
-
-enum class ReadStatus {
-    WATCHING, WATCHED, QUIT,
-}
