+++ app/src/main/java/com/teamwss/websoso/ui/novelInfo/model/ReadStatus.kt
@@ -1,0 +1,5 @@
+package com.teamwss.websoso.ui.novelInfo.model
+
+enum class ReadStatus {
+    WATCHING, WATCHED, QUIT,
+}
